Split up community page
